Tela de detalhes finalizada, Hard Coded, lint ok
