+++ src/components/RecipeDetails.js
@@ -6,6 +6,7 @@ import whiteHeartIcon from '../images/whiteHeartIcon.svg';
 import blackHeartIcon from '../images/blackHeartIcon.svg';
 import { getRecipe } from '../helpers/getFoodOrDrinkProperties';
 import Context from '../context/Context';
+import handleFavorite from '../helpers/handleFavorite';
 import '../styles/details.css';
 import { fetchRecipeById, fetchRecommendations } from '../services/api';
 
@@ -21,9 +22,9 @@ let RECIPE_DATA = {
   alcoholicOrNot: '',
 };
 
-const bodyWidth = document.body.offsetWidth;
-const LARGE_MOBILE_SCREEN = 425;
-const INIT_WIDTH = bodyWidth <= LARGE_MOBILE_SCREEN ? bodyWidth / 2 : '10rem';
+// const bodyWidth = document.body.offsetWidth;
+// const LARGE_MOBILE_SCREEN = 425;
+// const INIT_WIDTH = bodyWidth <= LARGE_MOBILE_SCREEN ? bodyWidth / 2 : '10rem';
 
 function start(callback, type, id) {
   callback(`/${type}s/${id}/in-progress`);
@@ -47,24 +48,19 @@ function storeRecipe(recipe, context) {
 }
 
 const RecipeDetails = () => {
-  const [width, setWidth] = useState(INIT_WIDTH);
+  const width = '250';
   const [object, setObject] = useState();
   const [featured, setFeatured] = useState();
   const [copied, setCopied] = useState(false);
   const [favorite, setFavorite] = useState(false);
-
   const history = useHistory();
   const { id } = useParams();
-
   const context = useContext(Context);
-
   const type = window.location.href.includes('comida') ? 'comida' : 'bebida';
   const featType = type === 'comida' ? 'bebida' : 'comida';
   const key = type === 'comida' ? 'meals' : 'cocktails';
-
   const { startedRecipes } = context;
   const hasStarted = startedRecipes[key].some((x) => x.id === id);
-
   if (object) RECIPE_DATA = { ...getRecipe(object, type) };
 
   useEffect(() => {
@@ -72,60 +68,37 @@ const RecipeDetails = () => {
     fetchRecommendations(type).then((arr) => setFeatured(arr));
   }, [type, id]);
 
-  window.addEventListener('resize', () => {
+  /*   window.addEventListener('resize', () => {
     const { offsetWidth } = document.body;
     if (offsetWidth <= LARGE_MOBILE_SCREEN) return setWidth(`${offsetWidth / 2}px`);
     setWidth('10rem');
-  });
+  }); */
 
   function handleShare() {
     navigator.clipboard.writeText(window.location.href);
     setCopied(true);
   }
 
-  function handleFavorite() {
-    const recipe = { id, ...FAV_RECIPE };
-    const localFavorites = localStorage.getItem('favoriteRecipes');
-    if (localFavorites) {
-      const localArray = JSON.parse(localFavorites);
-      const isFavorite = localArray.some((el) => (el.id === id));
-      if (isFavorite) {
-        const filteredArray = localArray.filter((el) => (el.id !== id));
-        localStorage.setItem('favoriteRecipes', JSON.stringify(filteredArray));
-        return setFavorite(!favorite);
-      }
-      const newFavorites = JSON.stringify([...localArray, recipe]);
-      localStorage.setItem('favoriteRecipes', newFavorites);
-      setFavorite(!favorite);
-    } else {
-      localStorage.setItem('favoriteRecipes', JSON.stringify([recipe]));
-      setFavorite(!favorite);
-    }
-  }
-
   let icon = whiteHeartIcon;
-
   const localFavorites = localStorage.getItem('favoriteRecipes');
 
   if (localFavorites) {
     const localArray = JSON.parse(localFavorites);
     const isFavorite = localArray.some((el) => (el.id === id));
-
     if (isFavorite) icon = blackHeartIcon;
   }
 
   function handleStart() {
     if (hasStarted) start(history.push, type, id);
-
     const recipe = { id, ...RECIPE_DATA };
-
     storeRecipe(recipe, context);
     start(history.push, type, id);
   }
 
   const { image, name, subCategory, instructions, ingredients, video } = RECIPE_DATA;
-  const { area, category, alcoholicOrNot} = RECIPE_DATA;
-  const FAV_RECIPE = {type, area, category, alcoholicOrNot, name, image};
+  const { area, category, alcoholicOrNot } = RECIPE_DATA;
+  const FAV_RECIPE = { type, area, category, alcoholicOrNot, name, image };
+  const paramsOfFavorite = { id, FAV_RECIPE, setFavorite, favorite };
 
   return (
     <div className="recipe-details">
@@ -158,7 +131,7 @@ const RecipeDetails = () => {
                 </button>
                 <button
                   type="button"
-                  onClick={ handleFavorite }
+                  onClick={ () => (handleFavorite(paramsOfFavorite)) }
                 >
                   <img
                     data-testid="favorite-btn"
@@ -167,19 +140,14 @@ const RecipeDetails = () => {
                     id={ id }
                   />
                 </button>
-
                 <br />
-
                 <h3
                   data-testid="recipe-category"
                 >
                   {subCategory}
                 </h3>
-
                 <h5>Ingredients</h5>
-
                 <br />
-
                 <ul>
                   {
                     ingredients.map((ingredient, index) => (
@@ -192,17 +160,13 @@ const RecipeDetails = () => {
                     ))
                   }
                 </ul>
-
                 <br />
-
                 <p
                   data-testid="instructions"
                 >
                   {instructions}
                 </p>
-
                 <br />
-
                 {
                   type === 'comida' && (
                     <iframe
@@ -212,13 +176,11 @@ const RecipeDetails = () => {
                     />
                   )
                 }
-
                 <div className="scroll-container">
                   <div className="recommendations">
                     {
                       featured && featured.map((value, index) => {
                         const feat = getRecipe(value, featType);
-
                         return (
                           <div
                             className="card"
@@ -237,7 +199,6 @@ const RecipeDetails = () => {
                     }
                   </div>
                 </div>
-
                 <button
                   type="button"
                   data-testid="start-recipe-btn"
@@ -246,7 +207,6 @@ const RecipeDetails = () => {
                 >
                   { hasStarted ? 'Continuar Receita' : 'Iniciar a Receita' }
                 </button>
-
               </div>
             </div>
           )
